requirements-dev.txt
# Include production requirements
-r requirements.txt

# Development & Testing Tools
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
Flask-Testing==0.8.1

# Code Quality & Linting
flake8==6.1.0
black==23.11.0
isort==5.12.0

# Development Server & Debugging
Flask-DebugToolbar==0.13.1

# Browser Testing
playwright==1.40.0
selenium==4.15.2

# Additional Development Utilities
python-dateutil==2.8.2
colorlog==6.8.0
watchdog==3.0.0

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Environment Management
python-decouple==3.8
